Increase the enumerated type of JdbcAuthType
